requirements.txt
pykafka==2.6.0
kafka==1.3.4.1
avro==1.8.2
sqlalchemy==1.1.14
redis-py-cluster==1.3.4
requests==2.11.1
behave==1.2.5
PyHamcrest==1.9.0
simplejson==3.10.0
pymongo==3.5.1
redis==2.10.5
junit2html==7
Fabric==1.14.0
etcd==2.0.8
python-etcd==0.4.3
